Переименовал Properties в Methods
