+++ src/test/java/order/GetOrderTest.java
@@ -8,7 +8,7 @@ import org.junit.Before;
 import org.junit.Test;
 import user.User;
 import user.UserChecks;
-import user.UserProperties;
+import user.UserMethods;
 
 import java.util.ArrayList;
 
@@ -17,8 +17,8 @@ public class GetOrderTest {
     User user;
     Order order;
     ArrayList<String> ingredients;
-    private final UserProperties userProperties = new UserProperties();
-    private final OrderProperties orderProperties = new OrderProperties();
+    private final UserMethods userMethods = new UserMethods();
+    private final OrderMethods orderMethods = new OrderMethods();
     private final UserChecks userChecks = new UserChecks();
     private final OrderChecks orderChecks = new OrderChecks();
 
@@ -26,7 +26,7 @@ public class GetOrderTest {
     @DisplayName("create new User")
     public void createUser() {
         user = User.getRandomUser();
-        ValidatableResponse response = userProperties.createNewUser(user);
+        ValidatableResponse response = userMethods.createNewUser(user);
         token = User.getToken(response);
     }
 
@@ -39,20 +39,20 @@ public class GetOrderTest {
         ingredients.add(Ingredient.getIngredients().get(1));
         ingredients.add(Ingredient.getIngredients().get(2));
         order.setIngredients(ingredients);
-        orderProperties.createOrder(order, token);
+        orderMethods.createOrder(order, token);
     }
 
     @Test
     @DisplayName("get Order with token")
     public void getOrderWithToken() {
-        ValidatableResponse response = orderProperties.getOrder(token);
+        ValidatableResponse response = orderMethods.getOrder(token);
         orderChecks.getOrderSuccessfully(response);
     }
 
     @Test
     @DisplayName("get Order without token")
     public void getOrderWithoutToken() {
-        ValidatableResponse response = orderProperties.getOrderWithoutToken();
+        ValidatableResponse response = orderMethods.getOrderWithoutToken();
         orderChecks.getOrderWithoutTokenUnSuccessfully(response);
     }
 
@@ -60,7 +60,7 @@ public class GetOrderTest {
     @DisplayName("delete User")
     public void deleteUser() {
         if (token!= null) {
-            ValidatableResponse response = userProperties.deleteExistingUser(token);
+            ValidatableResponse response = userMethods.deleteExistingUser(token);
             userChecks.deleteSuccessfully(response);
         }
     }
